+++ components/common/DropdownMenuContent/index.scss
@@ -4,6 +4,7 @@
   border-radius: 6px;
   box-shadow: 0 1px 4px rgb(0 0 0 / 8%);
   box-sizing: border-box;
+  overflow: auto;
   width: 30vw;
   max-width: 180px;
   margin: 0 $unit-2x;
@@ -130,6 +131,14 @@
     }
   }
 
+  & .destructive {
+    color: $error;
+
+    &:hover {
+      background: $error;
+      color: #fff;
+    }
+  }
   a {
     color: $grey-50;
 
@@ -177,12 +186,12 @@
 .MenuGroup {
   border-bottom: 1px solid var(--menu-separator);
 
-  &:first-child .MenuItem:first-child:hover {
+  &:first-child .MenuItem:first-child {
     border-top-left-radius: 6px;
     border-top-right-radius: 6px;
   }
 
-  &:last-child .MenuItem:last-child:hover {
+  &:last-child .MenuItem:last-child {
     border-bottom-left-radius: 6px;
     border-bottom-right-radius: 6px;
   }
